# docs/swagger.yaml
basePath: /
definitions:
  dto.CandidateResultsDTO:
    properties:
      candidate_id:
        type: string
      display_picture:
        type: string
      election_id:
        type: string
      name:
        type: string
      sex:
        type: integer
      user_id:
        type: string
      votes:
        type: integer
    type: object
  dto.CastVoteDTO:
    properties:
      candidate_id:
        type: string
      election_id:
        type: string
    required:
    - candidate_id
    - election_id
    type: object
...
  dto.CreateElectionDTO:
    properties:
      ending_at:
        type: string
      gender_specific:
        type: boolean
      locking_at:
        type: string
      starting_at:
        type: string
      title:
        type: string
    required:
    - ending_at
    - locking_at
    - starting_at
    - title
    type: object
...
  dto.DeleteParticipantDTO:
    properties:
      election_id:
        type: string
      participant_id:
        type: string
    required:
    - election_id
    - participant_id
    type: object
  dto.EditElectionDTO:
    properties:
      election_id:
        type: string
      ending_at:
        type: string
      gender_specific:
        type: boolean
      locking_at:
        type: string
      starting_at:
        type: string
      title:
        type: string
    required:
    - election_id
    type: object
  dto.Elections:
    properties:
      blacklisted:
        type: boolean
      election_id:
        type: string
      ending_at:
        type: string
      gender_specific:
        type: boolean
      locking_at:
        type: string
      starting_at:
        type: string
      title:
        type: string
      voted:
        type: boolean
    type: object
  dto.GeneralCandidateDTO:
    properties:
      approved:
        type: boolean
      candidate_id:
        type: string
      display_picture:
        type: string
      election_id:
        type: string
      first_name:
        type: string
      id_proof:
        type: string
      last_name:
        type: string
      poster:
        type: string
      register_no:
        type: string
      sex:
        type: integer
      user_id:
        type: string
    type: object
  dto.GeneralElectionDTO:
    properties:
      blacklisted:
        type: boolean
      candidate:
        $ref: '#/definitions/dto.GeneralCandidateDTO'
      candidates:
        items:
          $ref: '#/definitions/dto.GeneralCandidateDTO'
        type: array
      election_id:
        type: string
      ending_at:
        type: string
      gender_specific:
        type: boolean
      locking_at:
        type: string
      participants:
        items:
          $ref: '#/definitions/dto.GeneralParticipantDTO'
        type: array
      starting_at:
        type: string
      title:
        type: string
      voted:
        type: boolean
    type: object
  dto.GeneralElectionResultsDTO:
    properties:
      candidate_results:
        items:
          $ref: '#/definitions/dto.CandidateResultsDTO'
        type: array
      election_id:
        type: string
      ending_at:
        type: string
      fcandidate_results:
        items:
          $ref: '#/definitions/dto.CandidateResultsDTO'
        type: array
      gender_specific:
        type: boolean
      locking_at:
        type: string
      mcandidate_results:
        items:
          $ref: '#/definitions/dto.CandidateResultsDTO'
        type: array
      ocandidate_results:
        items:
          $ref: '#/definitions/dto.CandidateResultsDTO'
        type: array
      starting_at:
        type: string
      title:
        type: string
      total_participants:
        type: integer
      total_votes:
        type: integer
    type: object
  dto.GeneralParticipantDTO:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      participant_id:
        type: string
      register_number:
        type: string
      user_id:
        type: string
      voted:
        type: boolean
    type: object
...
  dto.Response:
    properties:
      message:
        type: string
    type: object
...
host: e1ect.herokuapp.com
info:
  contact:
    email: surajnm15@gmail.com
    name: ELECT API Support
  description: This is the backend server of ELECT web application.
  title: ELECT REST API
  version: "1.0"
paths:
  /api/candidate:
    post:
      operationId: enrollCandidate
      parameters:
      - in: formData
        name: election_id
        required: true
        type: string
      - in: formData
        name: sex
        required: true
        type: integer
      - description: Display Picture
        in: formData
        name: display_picture
        required: true
        type: file
      - description: poster
        in: formData
        name: poster
        required: true
        type: file
      - description: ID Proof
        in: formData
        name: id_proof
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Enroll as a candidate for the election you are part of
      tags:
      - candidate
  /api/candidate/approve/{id}:
    post:
      operationId: approveCandidate
      parameters:
      - description: Candidate ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Approve enrolled candidates to the election you created
      tags:
      - candidate
  /api/candidate/unapprove/{id}:
    post:
      operationId: unapproveCandidate
      parameters:
      - description: Candidate ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Unapprove enrolled candidates to the election you created
      tags:
      - candidate
  /api/election:
    post:
      operationId: election
      parameters:
      - description: Election Details
        in: body
        name: election
        required: true
        schema:
          $ref: '#/definitions/dto.CreateElectionDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Create Election if you are an Admin
      tags:
      - election
    put:
      operationId: election
      parameters:
      - description: Edit Election
        in: body
        name: election
        required: true
        schema:
          $ref: '#/definitions/dto.EditElectionDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Edit the election you created
      tags:
      - election
  /api/election/{id}:
    delete:
      operationId: electionP
      parameters:
      - description: Election ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Delete the election you created
      tags:
      - election
    get:
      operationId: electionP
      parameters:
      - description: Election ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GeneralElectionDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Get details of the election you created or you are part of
      tags:
      - election
  /api/elections:
    get:
      operationId: elections
      parameters:
      - description: Page
        in: query
        name: page
        type: string
      - description: Limit
        in: query
        name: limit
        type: string
      - description: Order By - starting_at
        in: query
        name: orderby
        type: string
      - description: Order - asc or desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.Elections'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Get a list of election you are part of OR you have created
      tags:
      - election
  /api/participant:
    delete:
      operationId: participant
      parameters:
      - description: Delete Participant
        in: body
        name: participant
        required: true
        schema:
          $ref: '#/definitions/dto.DeleteParticipantDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Delete the participant of the election you created
      tags:
      - participant
  /api/participants/{id}:
    post:
      operationId: addParticipants
      parameters:
      - description: Election ID
        in: path
        name: id
        required: true
        type: string
      - description: Participants List
        in: formData
        name: participants
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Add participants to the election you created
      tags:
      - participant
...
  /api/results/{id}:
    get:
      operationId: getElectionResults
      parameters:
      - description: Election ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GeneralElectionResultsDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Get the results of the election you were part of or you created
      tags:
      - election
  /api/vote:
    post:
      operationId: castVote
      parameters:
      - description: Cast Vote
        in: body
        name: vote
        required: true
        schema:
          $ref: '#/definitions/dto.CastVoteDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Cast vote to the candidate of the election you are part of
      tags:
      - participant
...
swagger: "2.0"